Refactor: port styles for ColonyTotalFunds
