--- src/components/common/ColonyHome/ColonyHome.css.d.ts
@@ -1,10 +1,0 @@
-export const main: string;
-export const mainContentGrid: string;
-export const leftAside: string;
-export const leftAsideNav: string;
-export const mainContent: string;
-export const domainsDropdownContainer: string;
-export const contentActionsPanel: string;
-export const rightAside: string;
-export const events: string;
-export const loadingWrapper: string;
