+++ src/components/common/ColonyTotalFunds/ColonyTotalFunds.css
@@ -4,6 +4,8 @@
 @value tokenDisplaySymbolFontSize: 18px;
 @value tokenDisplayFontWeight: 300;
 
+@value query700 from "~styles/queries.css";
+
 .main {
   display: block;
   margin-bottom: 32px;
@@ -82,4 +84,32 @@
 
 .selectedTokenSymbol:hover .tokenLockWrapper svg {
   fill: var(--colony-blue);
+}
+
+@media screen and query700 {
+  .main {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin: 0;
+    padding: 0 14px;
+    height: 77px;
+    border-bottom: 1px solid var(--temp-grey-13);
+  }
+
+  .totalBalanceCopy {
+    margin: 0;
+    font-size: var(--size-tiny);
+  }
+
+  .manageFundsLink {
+    display: flex;
+    align-items: center;
+    font-size: var(--size-smallish);
+  }
+
+  .selectedTokenAmount {
+    font-size: 24px;
+    line-height: 1;
+  }
 }
